+++ routes/index.js
@@ -15,12 +15,31 @@ router.get("/profile", (req, res, next) => {
     res.redirect("/");
     return;
   }
+
+  if (!req.user.verified) {
+    req.flash("error", "You must be approved by the admin to see this page");
+    res.redirect("/");
+    return;
+  }
+
   res.render("profile/profile.hbs");
 });
 
 
 //Delete profile
 router.get("/profile/:profileId/delete", (req, res,next)=> {
+  if (!req.user) {
+    req.flash("error", "You must be logged in to see this page");
+    res.redirect("/");
+    return;
+  }
+
+  if (!req.user.verified) {
+    req.flash("error", "You must be approved by the admin to see this page");
+    res.redirect("/");
+    return;
+  }
+  
   const { profileId } = req.params;
 
   User.findByIdAndRemove(profileId)
